// server/src/model/interface/iArtist.ts
export interface iArtist {
    id: number;
    estado: string;
    nombre: string;
    foto: Text;
    genero: string;
    descripcion: string;
    instrumentacion: string;
    telefono1: number;
    telefono2: number;
    email: string;
    web: string;
    redsocial1: string;
    redsocial2: string;
    redsocial3: string;
    notas: string;
    id_usuaria: number;
    id_ciudad: number;
  }